+++ tools/function-table.sh
@@ -12,9 +12,9 @@ NCDIR="$1"
 [ -d "$NCDIR" ] || { usage >&2 ; exit 1 ; }
 
 generate_lists () {
-  grep -h ^API "$1"/include/notcurses/*.h
-  grep -h -A1 ^static\ inline "$1"/include/notcurses/*.h | \
-    sed -e '/^--$/d' -e 'N;s/\n/ /' -e 's/\(.*\){$/\1;/'
+  grep -h ^API "$1"/include/notcurses/*.h | grep -v inline | sort
+  grep -h -A1 ^API\ inline "$1"/include/notcurses/*.h | \
+    sed -e '/^--$/d' -e 'N;s/\n/ /' -e 's/\(.*\){$/\1;/' | sort
 }
 
-generate_lists "$NCDIR" | sed -e 's/RESTRICT/restrict/g' | sort
+generate_lists "$NCDIR" | sed -e 's/RESTRICT/restrict/g'